Create README.md
This project includes the completion of 5 tasks related to graphic design. The tasks and their corresponding folders are as follows:

Task 1: Poster - This folder contains a poster design that can be used for advertising or promotional purposes.

Task 2: Logo - This folder contains a logo design that can be used for branding or identity purposes.

Task 3: Business Card - This folder contains a business card design that can be used for professional networking or contact information.

Task 4: Infographic - This folder contains an infographic design that can be used to visually represent information or data.

Task 5: Magazine - This folder contains a magazine design that can be used for print or digital publication.

All designs are original and created by me. Please feel free to use and share them.

Thank you for using Grafix Design Project!